+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Cthulhu 7 characters creating
 - rendering json data for PF2 characters
+- rendering skills for Cthulhu 7 characters
 
 ### Modified
 - rendering health for DH characters